course-refs - MCommunity URL as `actor` ID
Use the established MCommunity URL as the ID for the user in `actor`, etc.  Additionally, use only the URL instead of a whole `Person` object, as @arwhyte suggests (https://github.com/tl-its-umich-edu/udp/pull/12#discussion_r145497058).  This may imply that the `Person` may need to be defined earlier, using a Caliper "describe".